+++ scripts/project/test-jaarrekening-flow/run.sh
@@ -1,0 +1,2 @@
+#!/bin/bash
+exec node /script/main.js "$@"